Move payments to separate project
