+++ src/AgentFramework.AspNetCore/AgentFramework.AspNetCore.csproj
@@ -24,8 +24,7 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\AgentFramework.Core.Handlers\AgentFramework.Core.Handlers.csproj" />
-    <ProjectReference Include="..\AgentFramework.Core\AgentFramework.Core.csproj" />
+    <ProjectReference Include="..\AgentFramework\AgentFramework.csproj" />
   </ItemGroup>
 
 </Project>
