+++ src/AgentFramework.Core/AgentFramework.Core.csproj
@@ -15,6 +15,5 @@
   </ItemGroup>
   <ItemGroup>
     <Folder Include="Models\Payments\" />
-    <Folder Include="Decorators\Payments\" />
   </ItemGroup>
 </Project>
