+++ src/AgentFramework.Payments.SovrinToken/AgentFramework.Payments.SovrinToken.csproj
@@ -5,7 +5,8 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\AgentFramework.Core.Handlers\AgentFramework.Core.Handlers.csproj" />
+    <ProjectReference Include="..\AgentFramework.Payments\AgentFramework.Payments.csproj" />
+    <ProjectReference Include="..\AgentFramework\AgentFramework.csproj" />
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.Extensions.Hosting.Abstractions" Version="2.1.1" />
