+++ src/AgentFramework.Payments/AgentFramework.Payments.csproj
@@ -4,4 +4,13 @@
     <TargetFramework>netstandard2.0</TargetFramework>
   </PropertyGroup>
 
+  <ItemGroup>
+    <Folder Include="Abstractions\" />
+    <Folder Include="Records\" />
+    <Folder Include="Decorators\" />
+    <Folder Include="Middleware\" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\AgentFramework\AgentFramework.csproj" />
+  </ItemGroup>
 </Project>
