+++ src/AgentFramework.Payments/AgentMessageExtensions.cs
@@ -3,8 +3,10 @@ using AgentFramework.Core.Exceptions;
 using AgentFramework.Core.Messages;
 using AgentFramework.Core.Models.Payments;
 using AgentFramework.Core.Models.Records;
+using AgentFramework.Payments.Decorators;
+using AgentFramework.Payments.Records;
 
-namespace AgentFramework.Core.Decorators.Payments
+namespace AgentFramework.Payments
 {
     public static class AgentMessageExtensions
     {
