--- src/AgentFramework.Payments/Class1.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace AgentFramework.Payments
-{
-    public class Class1
-    {
-    }
-}
