+++ test/AgentFramework.Core.Tests/AgentFramework.Core.Tests.csproj
@@ -38,8 +38,6 @@
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\src\AgentFramework.AspNetCore\AgentFramework.AspNetCore.csproj" />
-    <ProjectReference Include="..\..\src\AgentFramework.Core.Handlers\AgentFramework.Core.Handlers.csproj" />
-    <ProjectReference Include="..\..\src\AgentFramework.Core\AgentFramework.Core.csproj" />
     <ProjectReference Include="..\..\src\AgentFramework.TestHarness\AgentFramework.TestHarness.csproj" />
     <ProjectReference Include="..\..\src\AgentFramework.Payments.SovrinToken\AgentFramework.Payments.SovrinToken.csproj" />
   </ItemGroup>
